fix 3612 * removing SGA associated classes, tools, shell scripts and tests, which are mostly skipped anyway * but keeps ExternalCommandlineProgramModule and (deprecated to prevent actual use) example on how to use it
